+++ meerschaum/utils/packages/__init__.py
@@ -11,13 +11,14 @@ import importlib, os, pathlib
 from meerschaum.utils.typing import Any, List, SuccessTuple, Optional, Union, Tuple, Dict
 from meerschaum.utils.threading import Lock, RLock
 from meerschaum.utils.packages._packages import packages, all_packages, get_install_names
+from meerschaum.utils.venv import (
+    activate_venv, deactivate_venv, venv_executable, venv_exec, venv_exists,
+    venv_target_path, inside_venv,
+)
 
 _import_module = importlib.import_module
 _import_hook_venv = None
-active_venvs = set()
 _locks = {
-    'active_venvs': RLock(),
-    'sys.path': RLock(),
     '_pkg_resources_get_distribution': RLock(),
 }
 _checked_for_updates = set()
@@ -280,6 +281,7 @@ def _install_to_import_name(install_name: str) -> str:
     """
     _install_no_version = get_install_no_version(install_name)
     return get_install_names().get(_install_no_version, _install_no_version)
+
 
 def get_install_no_version(install_name: str) -> str:
     """
@@ -607,287 +609,6 @@ def need_update(
     return False
 
 
-def is_venv_active(
-        venv: str = 'mrsm',
-        color : bool = True,
-        debug: bool = False
-    ) -> bool:
-    """
-    Check if a virtual environment is active.
-
-    Parameters
-    ----------
-    venv: str, default 'mrsm'
-        The virtual environment to check.
-
-    color: bool, default True
-        If `True`, include color in debug text.
-
-    debug: bool, default False
-        Verbosity toggle.
-
-    Returns
-    -------
-    A bool indicating whether the virtual environment `venv` is active.
-
-    """
-    if debug:
-        from meerschaum.utils.debug import dprint
-        dprint(f"Checking if virtual environment '{venv}' is active.", color=color)
-    return venv in active_venvs
-
-
-def deactivate_venv(
-        venv: str = 'mrsm',
-        color : bool = True,
-        debug: bool = False
-    ) -> bool:
-    """
-    Remove a virtual environment from `sys.path` (if it's been activated).
-
-    Parameters
-    ----------
-    venv: str, default 'mrsm'
-        The virtual environment to deactivate.
-
-    color: bool, default True
-        If `True`, include color in debug text.
-
-    debug: bool, default False
-        Verbosity toggle.
-
-    Returns
-    -------
-    Return a bool indicating whether the virtual environment was successfully deactivated.
-
-    """
-    import sys
-    if venv is None:
-        return True
-
-    if debug:
-        from meerschaum.utils.debug import dprint
-        dprint(f"Deactivating virtual environment '{venv}'...", color=color)
-
-    if venv in active_venvs:
-        with _locks['active_venvs']:
-            active_venvs.remove(venv)
-
-    if sys.path is None:
-        return False
-
-    target = venv_target_path(venv, debug=debug)
-    with _locks['sys.path']:
-        if str(target) in sys.path:
-            sys.path.remove(str(target))
-
-    if debug:
-        dprint(f'sys.path: {sys.path}', color=color)
-
-    return True
-
-
-tried_virtualenv = False
-def init_venv(venv: str='mrsm', debug: bool=False) -> bool:
-    """
-    Initialize the virtual environment.
-
-    Parameters
-    ----------
-    venv: str, default 'mrsm'
-        The name of the virtual environment to create.
-
-    Returns
-    -------
-    A `bool` indicating success.
-    """
-    if venv_exists(venv, debug=debug):
-        return True
-    import sys, platform, os, pathlib
-    from meerschaum.config._paths import VIRTENV_RESOURCES_PATH
-    global tried_virtualenv
-    try:
-        import venv as _venv
-        virtualenv = None
-    except ImportError:
-        _venv = None
-        virtualenv = None
-
-    venv_path = VIRTENV_RESOURCES_PATH / venv
-    bin_path = pathlib.Path(
-        venv_path,
-        ('bin' if platform.system() != 'Windows' else "Scripts")
-    )
-    _venv_success = False
-    if _venv is not None:
-        import io
-        from contextlib import redirect_stdout
-        f = io.StringIO()
-        with redirect_stdout(f):
-            _venv_success = run_python_package(
-                'venv',
-                [str(venv_path)] + (
-                    ['--symlinks'] if platform.system() != 'Windows' else []
-                ),
-                venv=None, debug=debug
-            ) == 0
-        if not _venv_success:
-            print("Please install python3-venv! Falling back to virtualenv...")
-        if not venv_exists(venv, debug=debug):
-            _venv = None
-    if not _venv_success:
-        virtualenv = attempt_import(
-            'virtualenv', venv=None, lazy=False, install=(not tried_virtualenv), warn=False,
-            check_update=False, color=False, debug=debug,
-        )
-        if virtualenv is None:
-            print(
-                "Failed to import `venv` or `virtualenv`! "
-                + "Please install `virtualenv` via pip then restart Meerschaum."
-            )
-            #  sys.exit(1)
-            return False
-
-        tried_virtualenv = True
-        try:
-            python_folder = (
-                'python' + str(sys.version_info.major) + '.' + str(sys.version_info.minor)
-            )
-            dist_packages_path = (
-                VIRTENV_RESOURCES_PATH /
-                venv / 'local' / 'lib' / python_folder / 'dist-packages'
-            )
-            local_bin_path = VIRTENV_RESOURCES_PATH / venv / 'local' / 'bin'
-            bin_path = VIRTENV_RESOURCES_PATH / venv / 'bin'
-            vtp = venv_target_path(venv=venv, allow_nonexistent=True, debug=debug)
-            virtualenv.cli_run([str(venv_path)])
-            if dist_packages_path.exists():
-                import shutil
-                shutil.move(dist_packages_path, vtp)
-                shutil.move(local_bin_path, bin_path)
-
-        except Exception as e:
-            import traceback
-            traceback.print_exc()
-            return False
-    return True
-
-
-def activate_venv(
-        venv: Optional[str] = 'mrsm',
-        color : bool = True,
-        debug: bool = False
-    ) -> bool:
-    """
-    Create a virtual environment (if it doesn't exist) and add it to `sys.path` if necessary.
-
-    Parameters
-    ----------
-    venv: Optional[str], default 'mrsm'
-        The virtual environment to activate.
-
-    color: bool, default True
-        If `True`, include color in debug text.
-
-    debug: bool, default False
-        Verbosity toggle.
-
-    Returns
-    -------
-    A bool indicating whether the virtual environment was successfully activated.
-
-    """
-    if venv in active_venvs:
-        return True
-    import sys, platform, os
-    from meerschaum.config._paths import VIRTENV_RESOURCES_PATH
-    if debug:
-        from meerschaum.utils.debug import dprint
-    #  old_cwd = pathlib.Path(os.getcwd())
-    #  os.chdir(VIRTENV_RESOURCES_PATH)
-    if venv is not None:
-        init_venv(venv=venv, debug=debug)
-    with _locks['active_venvs']:
-        if debug:
-            dprint(f"Activating virtual environment '{venv}'...", color=color)
-        active_venvs.add(venv)
-
-    target = venv_target_path(venv, debug=debug)
-    if str(target) not in sys.path:
-        sys.path.insert(0, str(target))
-    if debug:
-        dprint(f'sys.path: {sys.path}', color=color)
-    return True
-
-
-def venv_executable(venv: Optional[str] = 'mrsm') -> str:
-    """
-    The Python interpreter executable for a given virtual environment.
-    """
-    from meerschaum.config._paths import VIRTENV_RESOURCES_PATH
-    import sys, platform, os
-    return (
-        sys.executable if venv is None
-        else os.path.join(
-            VIRTENV_RESOURCES_PATH, venv, (
-                'bin' if platform.system() != 'Windows' else 'Scripts'
-            ), 'python'
-        )
-    )
-
-
-def venv_exec(
-        code: str,
-        venv: str = 'mrsm',
-        with_extras: bool = False,
-        as_proc: bool = False,
-        capture_output: bool = True,
-        debug: bool = False,
-    ) -> Union[bool, Tuple[int, bytes, bytes]]:
-    """
-    Execute Python code in a subprocess via a virtual environment's interpeter.
-    Return `True` if the code successfully executes, `False` on failure.
-
-    Parameters
-    ----------
-    code: str
-        The Python code to excecute.
-
-    venv: str, default 'mrsm'
-        The virtual environment to use to get the path for the Python executable.
-        If `venv` is `None`, use the default `sys.executable` path.
-
-    with_extras: bool, default False
-        If `True`, return a tuple of the exit code, stdout bytes, and stderr bytes.
-
-    as_proc: bool, default False
-        If `True`, return the `subprocess.Popen` object instead of executing.
-
-    Returns
-    -------
-    By default, return a bool indicating success.
-    If `as_proc` is `True`, return a `subprocess.Popen` object.
-    If `with_extras` is `True`, return a tuple of the exit code, stdout bytes, and stderr bytes.
-
-    """
-    import subprocess
-    from meerschaum.utils.process import run_process
-    from meerschaum.utils.debug import dprint
-    executable = venv_executable(venv=venv)
-    cmd_list = [executable, '-c', code]
-    if debug:
-        dprint(str(cmd_list))
-    if not with_extras and not as_proc:
-        return subprocess.call(cmd_list) == 0
-
-    stdout, stderr = (None, None) if not capture_output else (subprocess.PIPE, subprocess.PIPE)
-    process = subprocess.Popen(cmd_list, stdout=stdout, stderr=stderr)
-    if as_proc:
-        return process
-    stdout, stderr = process.communicate()
-    exit_code = process.returncode
-    return exit_code, stdout, stderr
-
 def get_pip(venv: Optional[str] = 'mrsm', debug: bool=False) -> bool:
     """
     Download and run the get-pip.py script.
@@ -916,12 +637,14 @@ def get_pip(venv: Optional[str] = 'mrsm', debug: bool=False) -> bool:
     cmd_list = [venv_executable(venv=venv), str(dest)] 
     return subprocess.call(cmd_list) == 0
 
+
 def pip_install(
-        *packages: List[str],
+        *packages: str,
         args: Optional[List[str]] = None,
+        requirements_file_path: Optional[pathlib.Path, str] = None,
         venv: Optional[str] = 'mrsm',
         deactivate: bool = True,
-        split : bool = False,
+        split: bool = False,
         check_update: bool = True,
         check_pypi: bool = True,
         check_wheel: bool = True,
@@ -941,6 +664,9 @@ def pip_install(
     args: Optional[List[str]], default None
         A list of command line arguments to pass to `pip`.
         If not provided, default to `['--upgrade']` if `_uninstall` is `False`, else `[]`.
+
+    requirements_file_path: Optional[pathlib.Path, str], default None
+        If provided, append `['-r', '/path/to/requirements.txt']` to `args`.
 
     venv: str, default 'mrsm'
         The virtual environment to install into.
@@ -1032,11 +758,12 @@ def pip_install(
                     color=False,
                 )
 
+    if requirements_file_path is not None:
+        _args.append('-r')
+        _args.append(str(pathlib.Path(requirements_file_path).resolve()))
+
     if not ANSI and '--no-color' not in _args:
         _args.append('--no-color')
-
-    if '--isolated' not in _args:
-        _args.append('--isolated')
 
     if '--no-input' not in _args:
         _args.append('--no-input')
@@ -1071,9 +798,15 @@ def pip_install(
             pass
 
     _packages = []
-    for p in packages:
-        root_name = p.split('.')[0] if split else p
-        install_name = _import_to_install_name(root_name)
+    for package_name in packages:
+        root_name = get_install_no_version(
+            package_name.split('.')[0] if split else package_name
+        )
+        import_name = get_install_names().get(root_name, root_name)
+        if root_name in get_install_names():
+            import_name = _install_to_import_name(root_name)
+
+        install_name = _import_to_install_name(import_name)
         if _uninstall:
             install_name = get_install_no_version(install_name)
         _packages.append(install_name)
@@ -1121,9 +854,19 @@ def completely_uninstall_package(
     removed from a virtual environment. 
     This is useful for dealing with multiple installed versions of a package.
     """
-    attempts, max_attempts = 0, 3
+    attempts = 0
+    _install_no_version = get_install_no_version(install_name)
+    clean_install_no_version = _install_no_version.lower().replace('-', '_')
+    installed_versions = []
+    for file_name in os.listdir(venv_target_path(venv, debug=debug)):
+        if not file_name.endswith('.dist-info'):
+            continue
+        clean_dist_info = file_name.replace('-', '_').lower()
+        if not clean_dist_info.startswith(clean_install_no_version):
+            continue
+        installed_versions.append(file_name)
+    max_attempts = len(installed_versions) + 1
     while attempts < max_attempts:
-        _install_no_version = get_install_no_version(install_name)
         if not venv_contains_package(
             _install_to_import_name(_install_no_version),
             venv=venv, debug=debug,
@@ -1308,7 +1051,6 @@ def attempt_import(
             print(f"Import hook for virtual environment '{_import_hook_venv}' is active.")
         venv = _import_hook_venv
 
-    _locks['sys.path'].acquire()
     activate_venv(venv=venv, color=color, debug=debug)
     _warnings = _import_module('meerschaum.utils.warnings')
     warn_function = _warnings.warn
@@ -1404,7 +1146,6 @@ def attempt_import(
         )
         modules.append(m)
 
-    _locks['sys.path'].release()
     if deactivate:
         deactivate_venv(venv=venv, debug=debug, color=color)
 
@@ -1432,6 +1173,7 @@ def lazy_import(
         **kw
     )
 
+
 def import_pandas(
         deactivate: bool = False,
         debug: bool = False,
@@ -1455,6 +1197,7 @@ def import_pandas(
         deactivate_venv(venv='mrsm', debug=debug)
     return pd
 
+
 def import_rich(
         lazy: bool = True, deactivate : bool = False, debug : bool = False,
         **kw : Any
@@ -1476,6 +1219,7 @@ def import_rich(
         deactivate_venv(venv='mrsm', debug=debug)
     return rich
 
+
 def _dash_less_than_2(**kw) -> bool:
     dash = attempt_import('dash')
     if dash is None:
@@ -1495,6 +1239,7 @@ def import_dcc(warn=False, **kw) -> 'ModuleType':
         attempt_import('dash_core_components', warn=warn, **kw)
         if _dash_less_than_2(warn=warn, **kw) else attempt_import('dash.dcc', warn=warn, **kw)
     )
+
 
 def import_html(warn=False, **kw) -> 'ModuleType':
     """
@@ -1559,6 +1304,7 @@ def get_modules_from_package(
         return _all, modules
 
     return modules
+
 
 def import_children(
         package: Optional['ModuleType'] = None,
@@ -1633,6 +1379,7 @@ def import_children(
     setattr(sys.modules[package_name], '__all__', _all)
     return members
 
+
 def reload_package(
         package: str,
         lazy: bool = False,
@@ -1652,6 +1399,7 @@ def reload_package(
             package_name = str(package)
     return pydoc.safeimport(package_name, forceload=1)
 
+
 def is_installed(
         import_name: str,
         venv: Optional[str] = 'mrsm',
@@ -1670,17 +1418,9 @@ def is_installed(
         found = importlib.util.find_spec(import_name) is not None
     except (ModuleNotFoundError, ValueError) as e:
         found = False
+    if deactivate:
+        deactivate_venv(venv=venv, debug=debug)
     return found
-
-    #  try:
-        #  mod = manually_import_module(import_name, split=False, venv=venv, debug=debug)
-    #  except ModuleNotFoundError:
-        #  mod = None
-    #  found = mod is not None and (venv == package_venv(mod))
-    #  if deactivate:
-        #  deactivate_venv(venv=venv)
-    #  return found
-    #  return venv_contains_package(import_name, venv=venv, debug=debug)
 
 
 def venv_contains_package(
@@ -1692,102 +1432,7 @@ def venv_contains_package(
     """
     Search the contents of a virtual environment for a package.
     """
-    #  if debug:
-        #  from meerschaum.utils.debug import dprint
     return get_module_path(import_name, venv=venv) is not None
-    #  vtp = venv_target_path(venv, debug=debug, allow_nonexistent=True)
-    #  if not vtp.exists():
-        #  return False
-    #  dir_name = _import_to_dir_name(
-        #  import_name.split('.')[0] if split else import_name
-    #  )
-    #  for file_name in os.listdir(vtp):
-        #  if file_name.endswith('dist-info') and file_name.startswith(dir_name):
-            #  if debug:
-                #  dprint(f"Found dist-info: {file_name}", color=False)
-            #  return True
-    #  return False
-
-
-def venv_exists(venv: Union[str, None], debug: bool = False) -> bool:
-    """
-    Determine whether a virtual environment has been created.
-    """
-    target_path = venv_target_path(venv, allow_nonexistent=True, debug=debug)
-    return target_path.exists()
-
-
-def venv_target_path(
-        venv: Union[str, None],
-        allow_nonexistent: bool = False,
-        debug: bool = False,
-    ) -> pathlib.Path:
-    """
-    Return a virtual environment's site-package path.
-
-    Parameters
-    ----------
-    venv: Union[str, None]
-        The virtual environment for which a path should be returned.
-
-    allow_nonexistent: bool, default False
-        If `True`, return a path even if it does not exist.
-
-    Returns
-    -------
-    The `pathlib.Path` object for the virtual environment's path.
-
-    """
-    import os, sys, platform, pathlib, site
-    from meerschaum.config._paths import VIRTENV_RESOURCES_PATH
-
-    ### Check sys.path for a user-writable site-packages directory.
-    if venv is None:
-        if not inside_venv():
-            return pathlib.Path(site.getusersitepackages())
-
-    venv_root_path = (
-        (VIRTENV_RESOURCES_PATH / venv)
-        if venv is not None else pathlib.Path(sys.prefix)
-    )
-    target_path = venv_root_path
-
-    ### Ensure 'lib' or 'Lib' exists.
-    lib = 'lib' if platform.system() != 'Windows' else 'Lib'
-    if not allow_nonexistent:
-        if not venv_root_path.exists() or lib not in os.listdir(venv_root_path):
-            print(f"Failed to find lib directory for virtual environment '{venv}'.")
-            import traceback
-            traceback.print_stack()
-            sys.exit(1)
-    target_path = target_path / lib
-
-    ### Check if a 'python3.x' folder exists.
-    python_folder = 'python' + str(sys.version_info.major) + '.' + str(sys.version_info.minor)
-    if target_path.exists():
-        target_path = (
-            (target_path / python_folder) if python_folder in os.listdir(target_path)
-            else target_path
-        )
-    else:
-        target_path = (
-            (target_path / python_folder) if platform.system() != 'Windows'
-            else target_path
-        )
-
-    ### Ensure 'site-packages' exists.
-    if allow_nonexistent or 'site-packages' in os.listdir(target_path): ### Windows
-        target_path = target_path / 'site-packages'
-    else:
-        import traceback
-        traceback.print_stack()
-        print(f"Failed to find site-packages directory for virtual environment '{venv}'.")
-        print("This may be because you are using a different Python version.")
-        print("Try deleting the following directory and restarting Meerschaum:")
-        print(VIRTENV_RESOURCES_PATH)
-        sys.exit(1)
-
-    return target_path
 
 
 def package_venv(package: 'ModuleType') -> Union[str, None]:
@@ -1800,18 +1445,6 @@ def package_venv(package: 'ModuleType') -> Union[str, None]:
         return None
     return package.__file__.split(str(VIRTENV_RESOURCES_PATH))[1].split(os.path.sep)[1]
 
-
-def inside_venv() -> bool:
-    """
-    Determine whether current Python interpreter is running inside a virtual environment.
-    """
-    import sys
-    return (
-        hasattr(sys, 'real_prefix') or (
-            hasattr(sys, 'base_prefix')
-                and sys.base_prefix != sys.prefix
-        )
-    )
 
 def ensure_readline() -> 'ModuleType':
     """Make sure that the `readline` package is able to be imported."""