+++ meerschaum/utils/sql.py
@@ -588,7 +588,7 @@ def update_query(
     target_table = get_sqlalchemy_table(target, connector)
     value_cols = []
     if debug:
-        dprint(f"{target_table.columns=}")
+        dprint(f"target_table.columns: {target_table.columns}")
     for c in target_table.columns:
         c_name, c_type = c.name, str(c.type)
         if c_name in join_cols:
@@ -597,7 +597,7 @@ def update_query(
             c_type = DB_FLAVORS_CAST_DTYPES[connector.flavor].get(c_type, c_type)
         value_cols.append((c_name, c_type))
     if debug:
-        dprint(f"{value_cols=}")
+        dprint(f"value_cols: {value_cols}")
 
     def sets_subquery(l_prefix: str, r_prefix: str):
         return 'SET ' + ',\n'.join([